Collisions are working! :O
WARNING : I changed entity positions : they are now topleft of the
Entity, not center. It's much easier for mot calculations.
If you want to get the center, you have to add half the width/height of
the Entity. (I updated the LightManager in consequence)

+++ src/net/watc4/game/entity/EntityLumi.java
@@ -10,7 +10,7 @@ import net.watc4.game.states.GameState;
 /** First Player : can fly and spreads light. */
 public class EntityLumi extends EntityPlayer
 {
-	public EntityLumi(int xPos, int yPos, GameState game)
+	public EntityLumi(float xPos, float yPos, GameState game)
 	{
 		super(xPos, yPos, game);
 		this.hasGravity = false;
